Deleted SelectPromotionButtons
